Add option "outer" per xpath to return entire node
By default, FeedIron returns the InnerHTML of selected nodes. Sometimes in creating a recipe it is easier to build a list of nodes to be kept instead of selecting the parent element and stripping out unwanted stuff in 'cleanup'. In these cases it may be helpful to have the entire selected node returned. This commit introduces a boolean option 'outer' per xpath (same as 'index'). If set to true, the entire node is returned. The verbose log shows whatever will be added to the result.